fix(tab-scroller): Cleaned up from review
