+++ packages/mdc-tab-scroller/foundation.js
@@ -22,9 +22,11 @@ import {
   strings,
 } from './constants';
 
+const INTERACTION_EVENTS = ['scroll', 'touchstart', 'pointerdown', 'mousedown', 'keydown'];
+
 /**
  * @extends {MDCFoundation<!MDCTabScrollerAdapter>}
- * @abstract
+ * @final
  */
 class MDCTabScrollerFoundation extends MDCFoundation {
   /** @return enum {string} */
@@ -59,24 +61,18 @@ class MDCTabScrollerFoundation extends MDCFoundation {
   /** @param {!MDCTabScrollerAdapter} adapter */
   constructor(adapter) {
     super(Object.assign(MDCTabScrollerFoundation.defaultAdapter, adapter));
-  }
 
-  /**
-   * Calculates a safe scroll value that is > 0 and < the max scroll value
-   * @param {number} scrollX The distance to scroll
-   * @return {number}
-   * @protected
-   */
-  calculateSafeScrollValue(scrollX) {
-    // Early exit for negative scroll values
-    if (scrollX < 0) {
-      return 0;
-    }
-    const contentWidth = this.adapter_.getContentOffsetWidth();
-    const rootWidth = this.adapter_.getOffsetWidth();
-    // Scroll values on most browsers are ints instead of floats so we round
-    const maxScrollValue = Math.round(contentWidth - rootWidth);
-    return Math.min(scrollX, maxScrollValue);
+    /** @private {function(?Event=)} */
+    this.handleInteraction_ = () => this.handleInteraction();
+
+    /** @private {function(?Event=)} */
+    this.handleTransitionEnd_ = () => this.handleTransitionEnd();
+
+    /** @private {boolean} */
+    this.shouldHandleInteraction_ = false;
+
+    /** @private {boolean} */
+    this.isAnimating_ = false;
   }
 
   /**
@@ -87,6 +83,53 @@ class MDCTabScrollerFoundation extends MDCFoundation {
     const currentTranslateX = this.calculateCurrentTranslateX_();
     const scrollLeft = this.adapter_.getScrollLeft();
     return scrollLeft - currentTranslateX;
+  }
+
+  /**
+   * Handles interaction events that occur during transition
+   */
+  handleInteraction() {
+    // Early exit if we're already handling the interaction
+    // This is necessary for cases like scrolling where the event handler can
+    // fire multiple times before it is unbound.
+    if (!this.shouldHandleInteraction_) {
+      return;
+    }
+
+    // Prevent other event listeners from handling this event
+    this.shouldHandleInteraction_ = false;
+    this.stopScrollAnimation_();
+    this.deregisterInteractionHandlers_();
+  }
+
+  /**
+   * Handles the transitionend event
+   */
+  handleTransitionEnd() {
+    this.isAnimating_ = false;
+    this.deregisterInteractionHandlers_();
+    this.adapter_.deregisterEventHandler('transitionend', this.handleTransitionEnd_);
+    this.adapter_.removeClass(MDCTabScrollerFoundation.cssClasses.ANIMATING);
+  }
+
+  /**
+   * Increment the scroll value by the given amount
+   * @param {number} scrollXIncrement The value by which to increment the scroll position
+   */
+  incrementScroll(scrollXIncrement) {
+    const currentScrollX = this.computeCurrentScrollPosition();
+    const safeScrollX = this.calculateSafeScrollValue_(currentScrollX + scrollXIncrement);
+    this.scrollTo_(safeScrollX, currentScrollX);
+  }
+
+  /**
+   * Scrolls to the given scrollX value
+   * @param {number} scrollX
+   */
+  scrollTo(scrollX) {
+    const currentScrollX = this.computeCurrentScrollPosition();
+    const safeScrollX = this.calculateSafeScrollValue_(scrollX);
+    this.scrollTo_(safeScrollX, currentScrollX);
   }
 
   /**
@@ -110,18 +153,102 @@ class MDCTabScrollerFoundation extends MDCFoundation {
   }
 
   /**
-   * Scrolls to the given scrollX value
-   * @param {number} scrollX
-   * @abstract
+   * Calculates a safe scroll value that is > 0 and < the max scroll value
+   * @param {number} scrollX The distance to scroll
+   * @return {number}
+   * @private
    */
-  scrollTo(scrollX) {} // eslint-disable-line no-unused-vars
+  calculateSafeScrollValue_(scrollX) {
+    // Early exit for negative scroll values
+    if (scrollX < 0) {
+      return 0;
+    }
+
+    const contentWidth = this.adapter_.getContentOffsetWidth();
+    const rootWidth = this.adapter_.getOffsetWidth();
+    // Scroll values on most browsers are ints instead of floats so we round
+    const maxScrollValue = Math.round(contentWidth - rootWidth);
+    return Math.min(scrollX, maxScrollValue);
+  }
 
   /**
-   * Increments the current scroll value by the given amount
-   * @param {number} scrollXIncrement
-   * @abstract
+   * Deregisters interaction events
+   * @private
    */
-  incrementScroll(scrollXIncrement) {} // eslint-disable-line no-unused-vars
+  deregisterInteractionHandlers_() {
+    INTERACTION_EVENTS.forEach((eventName) => {
+      this.adapter_.deregisterEventHandler(eventName, this.handleInteraction_);
+    });
+  }
+
+  /**
+   * Registers interaction events
+   * @private
+   */
+  registerInteractionHandlers_() {
+    INTERACTION_EVENTS.forEach((eventName) => {
+      this.adapter_.registerEventHandler(eventName, this.handleInteraction_);
+    });
+  }
+
+  /**
+   * Internal scroll method
+   * @param {number} newScrollX The new scroll position
+   * @param {number} currentScrollX The current scroll position
+   * @private
+   */
+  scrollTo_(newScrollX, currentScrollX) {
+    const scrollDelta = newScrollX - currentScrollX;
+    // Early exit if the scroll values are the same
+    if (scrollDelta === 0) {
+      return;
+    }
+
+    this.stopScrollAnimation_();
+    this.shouldHandleInteraction_ = false;
+
+    // This animation uses the FLIP approach.
+    // Read more here: https://aerotwist.com/blog/flip-your-animations/
+
+    this.adapter_.setContentStyleProperty('transform', `translateX(${scrollDelta}px)`);
+    // Setting scrollLeft triggers the repaint necessary for FLIP
+    this.adapter_.setScrollLeft(newScrollX);
+
+    requestAnimationFrame(() => {
+      this.adapter_.addClass(MDCTabScrollerFoundation.cssClasses.ANIMATING);
+      this.adapter_.setContentStyleProperty('transform', 'none');
+      this.isAnimating_ = true;
+      requestAnimationFrame(() => {
+        // Double-wrapped in a rAF because Firefox gets frisky with the scroll
+        // event and triggers a scroll event even though the handlers are bound
+        // *after* the scrollLeft value is set. This double-wrapped rAF ensures
+        // that we don't accidentally handle the scrollEvent we've triggered.
+        // See https://youtu.be/mmq-KVeO-uU?t=14m0s for a great explanation.
+        this.registerInteractionHandlers_();
+      });
+    });
+
+    this.adapter_.registerEventHandler('transitionend', this.handleTransitionEnd_);
+    this.shouldHandleInteraction_ = true;
+  }
+
+  /**
+   * Stops scroll animation
+   * @private
+   */
+  stopScrollAnimation_() {
+    // Early exit if not animating
+    if (!this.isAnimating_) {
+      return;
+    }
+
+    this.isAnimating_ = false;
+    const currentScrollPosition = this.computeCurrentScrollPosition();
+    this.adapter_.deregisterEventHandler('transitionend', this.handleTransitionEnd_);
+    this.adapter_.removeClass(MDCTabScrollerFoundation.cssClasses.ANIMATING);
+    this.adapter_.setContentStyleProperty('transform', 'translateX(0px)');
+    this.adapter_.setScrollLeft(currentScrollPosition);
+  }
 }
 
 export default MDCTabScrollerFoundation;
